Added more API docs
