+++ SDK/Nefarius.DsHidMini.IPC/Models/CmdStructs.cs
@@ -5,40 +5,65 @@ using Nefarius.DsHidMini.IPC.Models.Public;
 
 namespace Nefarius.DsHidMini.IPC.Models;
 
+/// <summary>
+///     What command is this message carrying
+/// </summary>
 [StructLayout(LayoutKind.Explicit)]
 [SuppressMessage("ReSharper", "InconsistentNaming")]
 internal struct DSHM_IPC_MSG_COMMAND
 {
+    /// <summary>
+    ///     Driver global command
+    /// </summary>
     [FieldOffset(0)]
     public DSHM_IPC_MSG_CMD_DRIVER Driver;
 
+    /// <summary>
+    ///     Device-specific command
+    /// </summary>
     [FieldOffset(0)]
     public DSHM_IPC_MSG_CMD_DEVICE Device;
 }
 
+/// <summary>
+///     Prefix of every packet describing the message
+/// </summary>
 [StructLayout(LayoutKind.Sequential)]
 [SuppressMessage("ReSharper", "InconsistentNaming")]
 [SuppressMessage("ReSharper", "UnusedMember.Global")]
 internal struct DSHM_IPC_MSG_HEADER
 {
-    // What request-behavior is expected (request, request-reply, ...)
+    /// <summary>
+    ///     What request-behavior is expected (request, request-reply, ...)
+    /// </summary>
     public DSHM_IPC_MSG_TYPE Type;
 
-    // What component is this message targeting (driver, device, ...)
+    /// <summary>
+    ///     What component is this message targeting (driver, device, ...)
+    /// </summary>
     public DSHM_IPC_MSG_TARGET Target;
 
-    // What command is this message carrying
+    /// <summary>
+    ///     What command is this message carrying
+    /// </summary>
     public DSHM_IPC_MSG_COMMAND Command;
 
-    // One-based index of which device is this message for
-    // Set to 0 if driver is targeted
+    /// <summary>
+    ///     One-based index of which device is this message for
+    /// </summary>
+    /// <remarks>Set to 0 if driver is targeted</remarks>
     public uint TargetIndex;
 
-    // The size of the entire message (header + payload) in bytes
-    // A size of 0 is invalid
+    /// <summary>
+    ///     The size of the entire message (header + payload) in bytes
+    /// </summary>
+    /// <remarks>A size of 0 is invalid</remarks>
     public uint Size;
 }
 
+/// <summary>
+///     Updates a specified devices' host address
+/// </summary>
 [StructLayout(LayoutKind.Sequential)]
 [SuppressMessage("ReSharper", "InconsistentNaming")]
 [SuppressMessage("ReSharper", "UnusedMember.Global")]
@@ -49,6 +74,9 @@ internal unsafe struct DSHM_IPC_MSG_PAIR_TO_REQUEST
     public fixed byte Address[6];
 }
 
+/// <summary>
+///     Reply to <see cref="DSHM_IPC_MSG_PAIR_TO_REQUEST" />.
+/// </summary>
 [StructLayout(LayoutKind.Sequential)]
 [SuppressMessage("ReSharper", "InconsistentNaming")]
 [SuppressMessage("ReSharper", "UnusedMember.Global")]
@@ -56,11 +84,20 @@ internal struct DSHM_IPC_MSG_PAIR_TO_REPLY
 {
     public DSHM_IPC_MSG_HEADER Header;
 
+    /// <summary>
+    ///     NTSTATUS of the set address action
+    /// </summary>
     public UInt32 WriteStatus;
-    
+
+    /// <summary>
+    ///     NTSTATUS of the get address action
+    /// </summary>
     public UInt32 ReadStatus;
 }
 
+/// <summary>
+///     Updates the player index of a given device
+/// </summary>
 [StructLayout(LayoutKind.Sequential)]
 [SuppressMessage("ReSharper", "InconsistentNaming")]
 [SuppressMessage("ReSharper", "UnusedMember.Global")]
@@ -68,9 +105,16 @@ internal struct DSHM_IPC_MSG_SET_PLAYER_INDEX_REQUEST
 {
     public DSHM_IPC_MSG_HEADER Header;
 
+    /// <summary>
+    ///     The new player index to set
+    /// </summary>
+    /// <remarks>Valid values are 1 to 7</remarks>
     public byte PlayerIndex;
 }
 
+/// <summary>
+///     Reply to <see cref="DSHM_IPC_MSG_SET_PLAYER_INDEX_REQUEST" />.
+/// </summary>
 [StructLayout(LayoutKind.Sequential)]
 [SuppressMessage("ReSharper", "InconsistentNaming")]
 [SuppressMessage("ReSharper", "UnusedMember.Global")]
@@ -81,22 +125,24 @@ internal struct DSHM_IPC_MSG_SET_PLAYER_INDEX_REPLY
     public UInt32 NtStatus;
 }
 
+/// <summary>
+///     Requests the driver host process PID and a wait handle for new input reports
+/// </summary>
+/// <remarks>Post-processing this command requires elevated privileges.</remarks>
 [SuppressMessage("ReSharper", "InconsistentNaming")]
 [StructLayout(LayoutKind.Sequential)]
 internal struct DSHM_IPC_MSG_GET_HID_WAIT_HANDLE_RESPONSE
 {
     public DSHM_IPC_MSG_HEADER Header;
 
+    /// <summary>
+    ///     The driver hosting process PID
+    /// </summary>
     public UInt32 ProcessId;
 
-    public IntPtr WaitHandle;  // HANDLE is typically represented as IntPtr in C#
-}
-
-[StructLayout(LayoutKind.Sequential)]
-[SuppressMessage("ReSharper", "InconsistentNaming")]
-internal struct IPC_HID_INPUT_REPORT_MESSAGE
-{
-    public UInt32 SlotIndex;
-
-    public DS3_RAW_INPUT_REPORT InputReport;
+    /// <summary>
+    ///     A handle to an auto-reset event
+    /// </summary>
+    /// <remarks>The requester of this handle must duplicate it into the current process before it becomes usable.</remarks>
+    public IntPtr WaitHandle;
 }
