+++ SDK/Nefarius.DsHidMini.IPC/Models/Enums.cs
@@ -2,54 +2,82 @@
 
 namespace Nefarius.DsHidMini.IPC.Models;
 
-// Describes the type of IPC message response behavior
+/// <summary>
+///     Describes the type of IPC message response behavior
+/// </summary>
 [SuppressMessage("ReSharper", "InconsistentNaming")]
 [SuppressMessage("ReSharper", "UnusedMember.Global")]
 internal enum DSHM_IPC_MSG_TYPE : UInt32
 {
-    // Invalid/reserved, do not use
+    /// <summary>
+    ///     Invalid/reserved, do not use
+    /// </summary>
     DSHM_IPC_MSG_TYPE_INVALID = 0,
 
-    // Client-to-driver data incoming, no acknowledgment/reply requested
+    /// <summary>
+    ///     Client-to-driver data incoming, no acknowledgment/reply requested
+    /// </summary>
     DSHM_IPC_MSG_TYPE_REQUEST_ONLY,
 
-    // Client-to-driver data incoming, must be acknowledged by reply
+    /// <summary>
+    ///     Client-to-driver data incoming, must be acknowledged by reply
+    /// </summary>
     DSHM_IPC_MSG_TYPE_REQUEST_RESPONSE,
 
-    // Client requested data, there is nothing to read for the driver
+    /// <summary>
+    ///     Client requested data, there is nothing to read for the driver
+    /// </summary>
     DSHM_IPC_MSG_TYPE_RESPONSE_ONLY,
 
-    // Driver-to-client response to a previous DSHM_IPC_MSG_TYPE_REQUEST_RESPONSE
+    /// <summary>
+    ///     Driver-to-client response to a previous DSHM_IPC_MSG_TYPE_REQUEST_RESPONSE
+    /// </summary>
     DSHM_IPC_MSG_TYPE_REQUEST_REPLY
 }
 
-// Describes the message receiver
+/// <summary>
+///     Describes the message receiver
+/// </summary>
 [SuppressMessage("ReSharper", "InconsistentNaming")]
 [SuppressMessage("ReSharper", "UnusedMember.Global")]
 internal enum DSHM_IPC_MSG_TARGET : UInt32
 {
-    // Invalid/reserved, do not use
+    /// <summary>
+    ///     Invalid/reserved, do not use
+    /// </summary>
     DSHM_IPC_MSG_TARGET_INVALID = 0,
 
-    // The message is targeted at the driver
+    /// <summary>
+    ///     The message is targeted at the driver
+    /// </summary>
     DSHM_IPC_MSG_TARGET_DRIVER,
 
-    // The message is targeted at a device
+    /// <summary>
+    ///     The message is targeted at a device
+    /// </summary>
     DSHM_IPC_MSG_TARGET_DEVICE,
 
-    // The message is targeted at the client/caller/app
+    /// <summary>
+    ///     The message is targeted at the client/caller/app
+    /// </summary>
     DSHM_IPC_MSG_TARGET_CLIENT
 }
 
-// Describes a per-driver command
+/// <summary>
+///     Describes a per-driver command
+/// </summary>
 [SuppressMessage("ReSharper", "InconsistentNaming")]
 [SuppressMessage("ReSharper", "UnusedMember.Global")]
 internal enum DSHM_IPC_MSG_CMD_DRIVER : UInt32
 {
-    // Invalid/reserved, do not use
+    /// <summary>
+    ///     Invalid/reserved, do not use
+    /// </summary>
     DSHM_IPC_MSG_CMD_DRIVER_INVALID = 0,
 
-    // Message without payload, useful to check for functionality
+    /// <summary>
+    ///     Message without payload, useful to check for functionality
+    /// </summary>
     DSHM_IPC_MSG_CMD_DRIVER_PING
 }
 
@@ -58,15 +86,23 @@ internal enum DSHM_IPC_MSG_CMD_DRIVER : UInt32
 [SuppressMessage("ReSharper", "UnusedMember.Global")]
 internal enum DSHM_IPC_MSG_CMD_DEVICE : UInt32
 {
-    // Invalid/reserved, do not use
+    /// <summary>
+    ///     Invalid/reserved, do not use
+    /// </summary>
     DSHM_IPC_MSG_CMD_DEVICE_INVALID = 0,
 
-    // Pair a given device to a new host
+    /// <summary>
+    ///     Pair a given device to a new host
+    /// </summary>
     DSHM_IPC_MSG_CMD_DEVICE_PAIR_TO,
-    
-    // Requests a player index update (switch player LED etc.)
+
+    /// <summary>
+    ///     Requests a player index update (switch player LED etc.)
+    /// </summary>
     DSHM_IPC_MSG_CMD_DEVICE_SET_PLAYER_INDEX,
-    
-    // Requests a wait handle for input report state changes
-    DSHM_IPC_MSG_CMD_DEVICE_GET_HID_WAIT_HANDLE,
+
+    /// <summary>
+    ///     Requests a wait handle for input report state changes
+    /// </summary>
+    DSHM_IPC_MSG_CMD_DEVICE_GET_HID_WAIT_HANDLE
 }
